Update codegen to latest spec

+++ src/main/java/io/swagger/configuration/BankApiApplicationRunner.java
@@ -9,6 +9,7 @@ import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -56,7 +57,8 @@ public class BankApiApplicationRunner implements ApplicationRunner
                 // Random amount between 10-3000
                 low = 10;
                 high = 3000;
-                Integer accountBalance = rand.nextInt(high-low) + low;
+                BigDecimal accountBalance = BigDecimal.valueOf(rand.nextInt(high-low) + low);
+
 
                 // Add IBAN and User to account
                 Iban iban = new Iban();
